+++ pandas/tests/window/common.py
@@ -1,0 +1,23 @@
+from datetime import datetime
+
+import numpy as np
+from numpy.random import randn
+
+from pandas import DataFrame, Series, bdate_range
+
+N, K = 100, 10
+
+
+class Base:
+
+    _nan_locs = np.arange(20, 40)
+    _inf_locs = np.array([])
+
+    def _create_data(self):
+        arr = randn(N)
+        arr[self._nan_locs] = np.NaN
+
+        self.arr = arr
+        self.rng = bdate_range(datetime(2009, 1, 1), periods=N)
+        self.series = Series(arr.copy(), index=self.rng)
+        self.frame = DataFrame(randn(N, K), index=self.rng, columns=np.arange(K))
